release: bug fixing
* fixes + test topics selection algo updates

* Squashed commit of the following:

commit 18422f445aa96b5ae3e87104380bc847295b9aeb
Author: Katerina Sidorenko <ekaterinasidorenko0914@gmail.com>
Date:   Wed Feb 18 21:56:42 2026 +0200

    add test docker compose

commit a8b6e933a8e9fef981c00b9e199291a16f33c111
Author: Katerina Sidorenko <ekaterinasidorenko0914@gmail.com>
Date:   Wed Feb 18 21:25:13 2026 +0200

    add test service dockerfile

commit 527a511b2119ca1360ba9b636b1dfaab8d401a5a
Author: Katerina Sidorenko <ekaterinasidorenko0914@gmail.com>
Date:   Tue Feb 17 21:16:14 2026 +0200

    add reference notes

commit 41197a47d641f841bd850265d8da1d8746a14ae9
Author: Katerina Sidorenko <ekaterinasidorenko0914@gmail.com>
Date:   Tue Feb 17 21:13:08 2026 +0200

    add docker file for core service

commit de970bee60d8a08311727f24d997a378efa1c8e3
Author: Katerina Sidorenko <ekaterinasidorenko0914@gmail.com>
Date:   Tue Feb 17 21:12:51 2026 +0200

    add util files

commit 0492b26432e031a5815527d7a3e797c85c01f43b
Author: Katerina Sidorenko <ekaterinasidorenko0914@gmail.com>
Date:   Tue Feb 17 19:52:49 2026 +0200

    add test data

commit 461c232640b39865aab8153a70137099f9ed3b2f
Author: Katerina Sidorenko <ekaterinasidorenko0914@gmail.com>
Date:   Sun Feb 15 10:59:36 2026 +0200

    set cover + layers

commit a3b153723bd4ca8880ff350b954957849f0f13c5
Author: Katerina Sidorenko <ekaterinasidorenko0914@gmail.com>
Date:   Sun Feb 15 10:45:58 2026 +0200

    set cover approach

commit 0e6ff327f1a4d03c796369acbe385737d776353c
Author: Katerina Sidorenko <ekaterinasidorenko0914@gmail.com>
Date:   Thu Feb 12 22:12:19 2026 +0200

    add deploy unit files

* refactor/core service (#28)

* update testing section styles

* just commit

* add feature folder

* add roadmap bookmarks

* naming convetions establish

* random updates

* on workspace rename

* add personal roadmap create event

* add simple workspace create logic

* delete all bullshit

* update persistence layer

* final persistance touches

* update

* workspace move

* add get personal roadmaps

* rename

* rename assesment

* add create personal roadmap

* create personal roadmap testing

* add get personal roadmaps

* add inbox task

* add snapshot create handler

* add author build snapshot of workspace

* add host service

* add inbox task migration + fixes

* add simple author roadmap build snapshot

* rename column

* add type create event

* add read roadmap blueprint methods

* rename folder to plural

* rename

* fix personal roadmaps filtering

* add workspaces mock methods

* add get workspaces

* add events models

* add events aggregation logic

* updates

* workspace snapshots update

* add aget personal roadmap

* update get workspace

* add title to worspace dto

* add apply of events in worspace

add apply events on get workspace

* add events processing

add delete connection

add unique items constraint

add connection fixes

add delete learning itme

update item test fixes

add publish event on workspace modification

add update learning item event

add events decorator layer + update add item logic

* fix workspace metadata conflict

* add get workspace summary

* fix get roadmap blueprint

* add blueptint roadmap snapshot

* add soft delete roadmap workspace

* add workspace title copy

* fix delete roadmap logic

* add progress and status calcualtion

* add progress and status calculation on snapshots

* Squashed commit of the following:

commit e6ca1c769703c25c9ff73797b21b1e8afdb5a250
Author: Katerina Sidorenko <ekaterinasidorenko0914@gmail.com>
Date:   Sat Mar 21 12:58:12 2026 +0200

    update author roadmap page

commit c22e3703ee4018d52a4559f7ec152d414f2cc533
Author: Katerina Sidorenko <ekaterinasidorenko0914@gmail.com>
Date:   Sat Mar 21 12:16:39 2026 +0200

    add publish roadmap

* fix image urls fatality

* styling updates

* asessment service fixes

fix assessments with copilot so who knows

rename testing services

* partial localization attempt

* attempt to fix assessment logic

* fixes + add mock dashboard

* Squashed commit of the following:

commit 1d05893e0f061f956e52f938b904ecbcfc1d8891
Author: Katerina Sidorenko <ekaterinasidorenko0914@gmail.com>
Date:   Thu Apr 2 20:15:25 2026 +0300

    Update index.tsx

commit c312ce076ef68b4eea46264da1af4ba81ccb0a2d
Author: Katerina Sidorenko <ekaterinasidorenko0914@gmail.com>
Date:   Thu Apr 2 16:15:13 2026 +0300

    add synchromization screen

commit f29a5298692ec9115783a538a053b14d22df2e6d
Author: Katerina Sidorenko <ekaterinasidorenko0914@gmail.com>
Date:   Wed Apr 1 21:12:16 2026 +0300

    fix cycle rejection

commit 76096e7a1dfa099cdee0cd1f76f0c3923d64b033
Author: Katerina Sidorenko <ekaterinasidorenko0914@gmail.com>
Date:   Wed Apr 1 20:13:38 2026 +0300

    add fe events queue

commit f6044c8c6d2e6d5474f25ebac76b0fe96733eb1d
Author: Katerina Sidorenko <ekaterinasidorenko0914@gmail.com>
Date:   Wed Apr 1 17:14:22 2026 +0300

    add be events amangement

commit 7a6ecefebc460eabfbc765cc5ea069dd203156f7
Author: Katerina Sidorenko <ekaterinasidorenko0914@gmail.com>
Date:   Wed Apr 1 16:40:13 2026 +0300

    update events models

commit 4ea1911a279ede56afa84265ae2dde1607386fdf
Author: Katerina Sidorenko <ekaterinasidorenko0914@gmail.com>
Date:   Wed Apr 1 16:22:22 2026 +0300

    move roadmap events types

commit af744252c26a9348f79f54d136e00bb9151c6ca7
Author: Katerina Sidorenko <ekaterinasidorenko0914@gmail.com>
Date:   Mon Mar 30 21:35:50 2026 +0300

    add roadmap validator

commit 2994594120665efe0dcc46fb3fc2ce6a5ed935b3
Author: Katerina Sidorenko <ekaterinasidorenko0914@gmail.com>
Date:   Mon Mar 30 21:20:30 2026 +0300

    extrcat editor component

commit b2df0c99eb68eb8697f0085ddbbcc405166231b5
Author: Katerina Sidorenko <ekaterinasidorenko0914@gmail.com>
Date:   Mon Mar 30 21:15:41 2026 +0300

    rename module

* add node type

* add cascade statuses updates

* update assessments flow

add learning items selection for intermidiate assessment

lots of assessment staff

assessment attempt create and get result

get assessemnt history

extract assessment hooks

update cascade logic updates

* remove unnessary files

* add simple seggestions algo

* add suggestion create and apply

* Update FrontierSubtopicSelector.cs

* feature: add initial testing

* bug: roadmap workspace progress recalculation (#31)

* Create 001-bug-roadmap-workspace-progress-recalculation.md

* add learning items projection table

* add learning items projection events

* add migrations

* renaming + add on snapshot updates

* fix on delete item

* fix calculation logic

* docs: introduce index table

* docs: editor problems statement + execution plan

* docs: add editor solution sequence diagram

* refactor: editor web sockets (#34)

* add web sockets constracts + hub

* move actions to websockets lib

* add action command dto

* temp

* move cycle check to bussiness

* add simple events processor

* add stream implementation

* add fe hub connection

* fixes

* fixes

* remove unused props

* remove prev implementation endpoints

* update localized resources

* remove complicated cache service

* build fixes

* fix workspace version consistency

* add errors localized toaster

* add on idempotency check

* add mock intermidiate snapshot taker

* bug: fix n+1 problem in workspaces

* refactor: add on workspace change taker

* bug: delete incident edges of node

* refactor: topic status recalculation

* refactor: assessment module

+++ backend/core-service/SkillMap.Business/RoadmapsWorkspace/IntegrationEvents/RoadmapWorkspaceChangedEvent.cs
@@ -3,8 +3,8 @@ using SkillMap.Shared.EventBus;
 
 namespace SkillMap.Business.RoadmapsWorkspace.IntegrationEvents;
 
-public record RoadmapWorkspaceChangedEvent(Guid Id, long WorkspaceId, long NewVersion, WorkspaceEventType EventType, DateTimeOffset OccurredDateTime) : IIntegrationEvent
+public record RoadmapWorkspaceChangedEvent(Guid Id, long WorkspaceId, int NewVersion, WorkspaceEventType EventType, DateTimeOffset OccurredDateTime) : IIntegrationEvent
 {
-    public static RoadmapWorkspaceChangedEvent Create(long workspaceId, long newVersion, WorkspaceEventType eventType)
+    public static RoadmapWorkspaceChangedEvent Create(long workspaceId, int newVersion, WorkspaceEventType eventType)
         => new RoadmapWorkspaceChangedEvent(Guid.NewGuid(), workspaceId, newVersion, eventType, DateTimeOffset.UtcNow);
 }
